example compositor: loop through xwayland surfaces
